requirements.txt
python-telegram-bot==20.7
SpeechRecognition==3.10.0
pydub==0.25.1
requests==2.31.0 